Combine rollup config into a unioned helper
