+++ package.json
@@ -46,6 +46,7 @@
     "read-yaml-promise": "^1.0.2",
     "rollup": "^0.55.1",
     "rollup-plugin-alias": "^1.4.0",
+    "rollup-plugin-auto-external": "^1.1.0",
     "rollup-plugin-babel": "^3.0.3",
     "rollup-plugin-commonjs": "^8.3.0",
     "rollup-plugin-json": "^2.3.0",
