+++ packages/slate-dev-logger/package.json
@@ -15,14 +15,7 @@
     "dist/umd/slate-dev-logger.min.js"
   ],
   "devDependencies": {
-    "mocha": "^2.5.3",
-    "rollup": "^0.55.1",
-    "rollup-plugin-babel": "^3.0.3",
-    "rollup-plugin-commonjs": "^8.3.0",
-    "rollup-plugin-node-resolve": "^3.0.2",
-    "rollup-plugin-replace": "^2.0.0",
-    "rollup-plugin-sourcemaps": "^0.4.2",
-    "rollup-plugin-uglify": "^3.0.0"
+    "mocha": "^2.5.3"
   },
   "scripts": {
     "build": "rollup --config",
