+++ packages/slate-plain-serializer/package.json
@@ -23,13 +23,6 @@
   },
   "devDependencies": {
     "mocha": "^2.5.3",
-    "rollup": "^0.55.1",
-    "rollup-plugin-auto-external": "^1.1.0",
-    "rollup-plugin-babel": "^3.0.3",
-    "rollup-plugin-commonjs": "^8.3.0",
-    "rollup-plugin-node-resolve": "^3.0.2",
-    "rollup-plugin-sourcemaps": "^0.4.2",
-    "rollup-plugin-uglify": "^3.0.0",
     "slate": "^0.32.2",
     "slate-hyperscript": "^0.5.2"
   },
