+++ packages/slate/package.json
@@ -30,15 +30,6 @@
     "react": "^0.14.0 || ^15.0.0 || ^16.0.0"
   },
   "devDependencies": {
-    "rollup": "^0.55.1",
-    "rollup-plugin-alias": "^1.4.0",
-    "rollup-plugin-auto-external": "^1.1.0",
-    "rollup-plugin-babel": "^3.0.3",
-    "rollup-plugin-commonjs": "^8.3.0",
-    "rollup-plugin-node-resolve": "^3.0.2",
-    "rollup-plugin-replace": "^2.0.0",
-    "rollup-plugin-sourcemaps": "^0.4.2",
-    "rollup-plugin-uglify": "^3.0.0",
     "slate-hyperscript": "^0.5.2"
   },
   "scripts": {
